// client/src/interfaces/User/UserFieldErrors.ts
export interface UserFieldErrors {
    first_name?: string[];
    last_name?: string[];
    user_name?: string[];
    user_email?: string[];
    password_confirmation?: string[];
    password?: string[];
    user_phone?: string[];
    user_address?: string[];
    user_image?: string[];
    role?: string[];
  }